Add a key state management unit
The need for this unit is prompted by the needs of AES inverse
functions: our AES inverse implementation needs the last AES round key
as an input, and computing the key schedule is not something we want to
delegate to the user. Further, this last round key should be stored,
to avoid re-computing the AES key schedule before each block processing
(that would double the latency).

The new key holder also manages the key size and direction (enc/dec).
Its interface is more serial: a single 32-bit bus, which will make it
easy to integrate with a system bus without the need for an additional
buffer.